backend/requirements.txt
# Core
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.1
Flask-JWT-Extended==4.6.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30

# Utils
python-dotenv==1.0.1
Werkzeug==3.1.3

# Computer Vision & Automation
opencv-python
mediapipe==0.10.9
numpy
PyAutoGUI

# AI & Hub
huggingface_hub
groq
requests
